# Conflicts:
#   server/src/controllers/bookController.js
#   server/src/controllers/userController.js
#   server/src/index.js
#   server/src/routes/bookRouter.js
#   server/src/routes/userRouter.js

+++ server/src/controllers/userController.js
@@ -1,5 +1,6 @@
 import bcrypt from "bcrypt";
 import UserModel from "../models/userModel.js";
+import BookModel from "../models/bookModel.js";
 import { GetBookByID as GetBookByIsbn } from "./bookController.js";
 
 export const newUser = async (req, res) => {
@@ -31,48 +32,72 @@ export const newUser = async (req, res) => {
         res.status(404).end();
     }
 };
-
+async function GetAndValidateRequestingUser(req) {
+    const {
+        id, password,
+    } = req.body;
+    const requester = await UserModel.findOne({ id }).exec();
+    if(bcrypt.compareSync(password, requester.password))
+        return requester;
+    return undefined;
+}
 export async function GetUserOrFail(req, res) {
-    const user = await UserModel.findOne(req.body.filter).exec();
-    if (user) {
-        res.status(200).json(user);
+    const requester = await GetAndValidateRequestingUser(req);
+    if (requester){
+        const user = await UserModel.findOne(req.body.filter).exec();
+        //Only allow normal users to seach themselves.
+        if (user && (user.id === requester.id || requester.role === "admin")){
+            res.status(200).json(user);
+        } else {
+            res.status(400).json({ Error: "NotFound" });
+        }
     } else {
         res.status(400).json({ Error: "NotFound" });
     }
 }
 export async function GetAllUsersOrFail(req, res) {
-    const allUsers = await UserModel.find().exec();
-    if (allUsers) {
-        res.status(200).json(allUsers);
+    const requester = await GetAndValidateRequestingUser(req);
+    if (requester.role === "admin") {
+        const allUsers = await UserModel.find().exec();
+        if (allUsers) {
+            res.status(200).json(allUsers);
+        } else {
+            res.status(400).json({ Error: "NotFound" });
+        }
     } else {
         res.status(400).json({ Error: "NotFound" });
     }
 }
 export async function DeleteUserOrFail(req, res) {
-    const {
-        id,
-    } = req.body;
-    const filter = {
-        id,
-    };
-    const user = await UserModel.findOneAndRemove(
-        filter,
-        { useFindAndModify: false },
-    );
-    if (user) {
-        res.status(200).json(user);
+    const requester = await GetAndValidateRequestingUser(req);
+    if (requester.role === "admin") {
+        const {
+            id,
+        } = req.body;
+        const filter = {
+            id,
+        };
+        const user = await UserModel.findOneAndRemove(
+            filter,
+            { useFindAndModify: false },
+        );
+        if (user) {
+            res.status(200).json(user);
+        } else {
+            res.status(400).json({ Error: "NotFound" });
+        }
     } else {
         res.status(400).json({ Error: "NotFound" });
     }
 }
 export async function ModifyUserOrFail(req, res) {
-    const {
-        id,
-        password,
-    } = req.body;
-    const filter = { id, password };
-    const account = await UserModel.findOne(filter).exec();
-    if (account) {
+
+    const requester = await GetAndValidateRequestingUser(req);
+    const password  = req.body;
+    const id = req.body.replacementData.id;
+    const account = (id) ? await UserModel.findOne(id).exec() : requester;
+    if (account && (bcrypt.compareSync(password, account.password) ||
+                    requester.role === "admin")) {
         const rd = req.body.replacementData;
 
         const name = rd.name ? rd.name : account.name;
@@ -151,3 +176,36 @@ export const userBorrowBook = async (req, res) => {
         res.status(404).end();
     }
 };
+//Maybe this should be part of bookController...
+export async function ReserveBookForUserOrFail(req, res) {
+    const user = await GetAndValidateRequestingUser(req);
+    if (user) {
+        const isbn = req.body.isbn;
+        const copy = Number(req.body.copy);
+        const book = await GetBookByIsbn(isbn);
+        if (book) {
+            const bookCopy = book.copies.find((it) => it.id === copy);
+            if (bookCopy) {
+                if (!bookCopy.reserveList.find( it => it.reserveId === user.id)) {
+                    bookCopy.reserveList = bookCopy.reserveList.concat(
+                        { reserveId: user.id });
+                    
+                    await BookModel.findOneAndUpdate(
+                        { isbn },
+                        book,
+                        { useFindAndModify: false, new: true },
+                    ).exec();
+                    res.status(200).json(book);
+                } else{
+                    res.status(400).json({ Error: "AlreadyOnReserveList" });
+                }
+            } else {
+                res.status(400).json({ Error: "CopyNotFound" });
+            }
+        } else {
+            res.status(400).json({ Error: "BookNotFound" });
+        }
+    } else {
+        res.status(400).json({ Error: "UserNotFound" });
+    }    
+}
